// dashboard/src/api/models/creditNoteLine.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { CreditNoteLineTax } from "./creditNoteLineTax";

export interface CreditNoteLine {
  id: string;
  /** @nullable */
  invoice_line_id: string | null;
  description: string;
  /** @nullable */
  quantity: number | null;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  unit_amount: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  amount: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  total_amount_excluding_tax: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  total_tax_amount: string;
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  total_amount: string;
  readonly taxes: readonly CreditNoteLineTax[];
}